backend/requirements.txt
asgiref==3.8.1
Django
django-cors-headers==4.6.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg-pool==3.2.4
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
uvicorn
aiogram>=3.0
aiohttp>=3.8